# src/api/swaggerV6.yaml
swagger: '2.0'
info:
  description: Governify Registry
  version: '6'
  title: Governify Registry
  termsOfService: www.isa.us.es
  contact:
    name: ISA Research Group
    url: www.isa.us.es
    email: isagroup.us@gmail.com
  license:
    name: MIT
    url: 'http://opensource.org/licenses/MIT'
host: registry.governify.io
basePath: /api/v6
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/states/{agreement}/guarantees':
    get:
      tags:
        - state-registry
      description: XXX
      operationId: getAllGuarantees
      parameters:
        - name: agreement
          in: path
          description: XXX
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/guarantee_state'
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/error_model'
      x-swagger-router-controller: StateController
...
definitions:
...
  scope:
    type: object
  log:
    type: object
...
  guarantee_state:
    type: object
    properties:
      guarantee:
        type: string
      scope:
        $ref: '#/definitions/scope'
      logs:
        $ref: '#/definitions/log'
      period:
        $ref: '#/definitions/period'
      evidences:
        type: array
        items:
          $ref: '#/definitions/evidence_state'
      fulfilled:
        type: boolean
  period:
    type: object
    properties:
      from:
        type: string
      to:
        type: string
...
  evidence_state:
    type: object
    properties:
      evidence:
        type: boolean
      issue_init:
        type: string
      issue_end:
        type: string
      issue_duration:
        type: string
...
  error_model:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
...